Pass chained layer digests to stargz-based snapshotters for parallel preparation
Currently, CRI plugin passes each layer digest to remote snapshotters
sequentially, which leads to sequential snapshots preparation. But it costs
extra time for remote snapshotters which need to connect to the remote backend
store (e.g. registries) for checking the snapshot existence on each preparation.

This commit solves this problem by passing all layer digests in an image to
snapshotters and by allowing them to prepare these snapshots in parallel, which
leads to speed up the preparation.

Signed-off-by: Koehi Tokunaga <ktokunaga.mail@gmail.com>